+++ src/main/ruby/core/posix.rb
@@ -244,7 +244,7 @@ module Truffle::POSIX
     attach_function :setpriority, [:int, :id_t, :int], :int
 
     attach_function :execve, [:string, :pointer, :pointer], :int
-    attach_function :truffleposix_posix_spawnp, [:string, :pointer, :pointer, :int, :pointer, :int], :pid_t, LIBTRUFFLEPOSIX
+    attach_function :truffleposix_posix_spawnp, [:string, :pointer, :pointer, :int, :pointer, :int, :int, :pointer], :pid_t, LIBTRUFFLEPOSIX
     attach_function :truffleposix_waitpid, [:pid_t, :int, :pointer], :pid_t, LIBTRUFFLEPOSIX, true
 
     # ENV-related